Fix docker compile for wsl2
Testing with wsl2 were presenting failure when
compiling with docker due to the 'cwd' process folder.
The code was taking the VSCode process dir instead, which
was a windows path unrelated to the workspace. With this
modification, the container can properly find the contract
to be compiled, using the proper path.